+++ src/router/pawtens/Pawtens_detail.js
@@ -1,6 +1,3 @@
-import { useState } from "react";
-import axios from "axios";
-
 function Pawtens_detail(){
     return (
         <div>
