+++ src/styles/search.css
@@ -1,0 +1,24 @@
+.pawtens {
+    width: 1200px;
+    float: left;
+}
+.pawtensList {
+    margin-top: 30px;
+    display: flex;
+    flex-direction: row;
+    flex-wrap: wrap;
+    align-content: flex-start;
+}
+.pawtensItem {
+    margin: 10px;
+}
+
+.pawtensItem img {
+    width:250px;
+    height:180px;
+    object-fit: cover;
+}
+
+.createPawtens .filePreview {
+    margin-top: 20px;
+}
